Minor fixes

+++ src/dynohook/decoder.cpp
@@ -479,8 +479,8 @@ std::vector<uint8_t*> Decoder::findRelativeInstructionsOfType(void* startAddress
  */
 bool Decoder::calculateRipRelativeMemoryAccessBounds(void* sourceAddress, size_t length, int64_t& lowestAddress, int64_t& highestAddress) const {
     size_t byteCount = 0;
-    int64_t tmpLowestAddress = (int64_t) 0xffffffffffffffff;
-    int64_t tmpHighestAddress = 0;
+    uint64_t tmpLowestAddress = 0xffffffffffffffff;
+    uint64_t tmpHighestAddress = 0;
 
     // we will atleast relocate "length" bytes. To avoid splitting an instruction we might relocate more.
     while (byteCount < length) {
